+++ src/components/mainPage/about-section.css
