+++ src/components/mainPage/example-section.css
