+++ src/components/mainPage/mainpage.css
