+++ src/components/mainPage/upload.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'Inter';
-  src: local('Inter'), url(../fonts/Inter/Inter-VariableFont_slnt,wght.ttf) format('truetype');
+  src: local('Inter'), url(../../fonts/Inter/Inter-VariableFont_slnt,wght.ttf) format('truetype');
   font-weight: 100 1000;
   font-stretch: 25% 151%;
 }
